requirements.txt
ultralytics==8.3.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23.0
opencv-python>=4.9.0
pillow>=10.0.0
matplotlib>=3.8.0
streamlit>=1.36.0
onnx>=1.16.0
onnxruntime>=1.17.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0.1
geopandas>=0.14.0
sentinelhub==3.10.0